Few tweaks trying to run the live process
I've hit two major errors that have stopped the process of finalizing the GAP species DOIs. One seemed to be a timeout problem. I restarted with a check on public vs. reserved, and it continued until hitting a key error on "doi." That doesn't make any sense, but it's currently stuck at that point.